main_test.goの削除対応

+++ shell/save.sh
@@ -1,6 +1,13 @@
 #!/bin/bash
 . shell/contests_function.sh
 
+# 雑に引数があったら削除しない
+if [ $# = 1 ]; then
+  remove=false
+else
+  remove=true
+fi
+
 problem_url=`get_problem_url`
 
 problem_directory=`get_problem_directory`
@@ -14,5 +21,8 @@ echo ${problem_url} > ${problem_directory}/README.md
 git add ${problem_directory}
 git commit -m "`get_current site` `get_current title` `get_current problem`"
 
-rm main.go
-rm -rf testdata
+if "${remove}"; then
+  rm main.go
+  rm main_test.go
+  rm -rf testdata
+fi
